First attempt to fix Windows dependencies
Windows does not support grp and pwd modules, which this code uses to
provide an ls-like listing. Trying an approach with the platform
module to see if it works to not import and not use the modules. This
may cause other problems and also PEP-8 issues, hence af first attempt.