Changes to support citations better.
Moved the instance variable `bibliography` to `_bibliography` so that
it is not written to the flowchart, and upgraded `to_dict()` to ignore
all but specified attributes starting with and underscore. `_uuid` is
the main exception.

Finally, in order to facilitate debugging, added an optional logger to
the initializer and changed all logger commands to using the instance
variable. By default it is still the file-local logger, but this
allows subclasses to pass in their logger, controlled by a command
line option such as --packmol-log-level to allow more precise
debugging.